Bumped the change log file a bit to make version 5.3 only having one change. Version 5.1 now has some of the changes that version 5.2 had. Both versions will now have 10 equal changes.
Versions 5.1 to 5.2 have the folowing changes.
Revampt the intire readme with added and removed keyboard shortcuts. I also took out large amounts of Unnecessary info that didn't need to be there in preparation for a more thorough documentation of the game's mecanics.
Lives no longer reset themselves to having only 1 of them after death.
Optimized the slider class and made it cleaner to read.
Lockable doors and passages can now be spied with the spire.
You can now build lockable passages with a password in the builder menu. They will also be written to the map you're currently working on.
You can now build lockable doors with a password in the builder menu. They will also be written to the map you're currently working on.
Fixed all of the npc related zones to only trigger them if you are at the exact y axes at witch they were built.
Fixed a bug with the slider class returning invalid entries when ran for the first time.
Completely eliminated the need to type values for building a reverb effect in the builder menu. Instead of inputting the value in for how much you want the Parameter to be at, you can quickly asine the value by using the arrow keys on the keyboard and pressing tab to focus on the value you want to change. Thanks to the newly created slider class that will handel all of this with a breez. You can also press the home and end keys to directly jump to the minimum and maximum values of a Parameter you're curren'tly focused on.
Started working on slider support for adjusting effect parameters.
Fixed a bug where hazards would spawn random ledge platforms on both sides that were different than the ones spesified on the left side of the hazard. If you build a hazard and set the ledge spawning to true, it will spawn random ledge platforms for both sides of the hazard, but keep it consistent with the ledge platform that has been spawned on the left side of the hazard.
added back the wall that you can bump into when making an instrument in the builder menu.
Added back the ability to spawn ledges in the builder prompt while building hazards on the map.
Bomb zones no longer have minimum and maximum y values when building them in the builder menu.
Npc zones such as animals, enemies, and projectiles no longer have minimum and maximum y values when building them in the builder menu.
Projectiles can now damage anything in it's range, not just you.
You can now build music ambiences in the builder menu. They will also be written to the map you're currently working on.
renamed music to music source in the builder menu.
Renamed ambience to sound ambience in the builder menu.