Made some major enhancements to all sliders. The volume and pitch Parameters no longer take the virtual input box. It now uses sliders to adjust both of them without relying on the user to type in the values.
Made it so objects that take the volume and pitch parameter use sliders to adjust them instead of the vertual input box asking the user to type in the values they want. This change also applies to the other game modes as well, not just the builder mode.The volume and pitch sliders now anounce values in percentages, not actual numbers. The reverb sliders will remain largely unchanged, aside from a few tweaks here and their.