Fixed lots of bugs related to lockable doors and teleporters.
Teleporters can now be spied with the spire.
Added a prompt to the teleporter builder that asks weather you'd like the teleporter to be singgle use or not.
Added lockable doors to the object info menu.
Lockable doors are now destroyable just like regular doors.
Fixed the bug of not being able to move up or down on the Y axes when a non staircase platform was detected on the map.