Added forcefields to the builder menu and spire, along with other bug fixes.
Fixed various bugs with forcefields and vanishing platforms.
Forcefields can now be spied with the spire. When you use the spire on one of them, it will tell you weather it's activated or deactivated, along with it's asined length and width.
You can now build forcefields in the builder menu. They will also be written to the map you're currently working on.