Added mines to the builder menu and spire, along with fixing several issues with sounds not destroying or updating properly in various classes.
Addressed lots of issues related to sounds not destroying or updating properly when they should.
Mines can now damage anything in it's range, not just you.
Mines can now be spied with the spire. When you use the spire on one of them, it will tell you weather it's active or not.
You can now build mines in the builder menu. They will also be written to the map you're currently working on.